removed duplicate files for sqlite providers
