+++ src/ServiceStack.OrmLite.Sqlite32/ServiceStack.OrmLite.Sqlite32.csproj
@@ -47,8 +47,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\ServiceStack.OrmLite.Sqlite\SqliteExpressionVisitor.cs">
+      <Link>SqliteExpressionVisitor.cs</Link>
+    </Compile>
+    <Compile Include="..\ServiceStack.OrmLite.Sqlite\SqliteOrmLiteDialectProviderBase.cs">
+      <Link>SqliteOrmLiteDialectProviderBase.cs</Link>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SqliteExpressionVisitor.cs" />
     <Compile Include="SqliteOrmLiteDialectProvider.cs" />
   </ItemGroup>
   <ItemGroup>
