+++ src/ServiceStack.OrmLite.Sqlite64/ServiceStack.OrmLite.Sqlite64.csproj
@@ -46,8 +46,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\ServiceStack.OrmLite.Sqlite\SqliteExpressionVisitor.cs">
+      <Link>SqliteExpressionVisitor.cs</Link>
+    </Compile>
+    <Compile Include="..\ServiceStack.OrmLite.Sqlite\SqliteOrmLiteDialectProviderBase.cs">
+      <Link>SqliteOrmLiteDialectProviderBase.cs</Link>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SqliteExpressionVisitor.cs" />
     <Compile Include="SqliteOrmLiteDialectProvider.cs" />
   </ItemGroup>
   <ItemGroup>
